+++ go/internal/website/vulnerability_view.go
@@ -1,0 +1,403 @@
+package website
+
+import (
+	"fmt"
+	"html/template"
+	"log/slog"
+	"net/url"
+	"slices"
+	"strings"
+	"time"
+
+	"github.com/google/osv.dev/go/logger"
+	"github.com/ossf/osv-schema/bindings/go/osvschema"
+	gocvss20 "github.com/pandatix/go-cvss/20"
+	gocvss30 "github.com/pandatix/go-cvss/30"
+	gocvss31 "github.com/pandatix/go-cvss/31"
+	gocvss40 "github.com/pandatix/go-cvss/40"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
+)
+
+// Link returns the Git commit link for an event if applicable.
+func (e AffectedEventDisplay) Link() string {
+	return GitCommitLink(e.Repo, e.Value)
+}
+
+// DepsDevURL generates the deps.dev insights URL for the affected package.
+func (p AffectedPackageDisplay) DepsDevURL() string {
+	if p.EcosystemName == "" || p.PackageName == "" {
+		return ""
+	}
+
+	baseEco, _, _ := strings.Cut(p.EcosystemName, ":")
+	system, exists := osvToDepsDevEcosystemsMap[baseEco]
+	if !exists {
+		return ""
+	}
+
+	pkgName := p.PackageName
+	// Normalize PyPI package names if ecosystem is PyPI
+	if baseEco == "PyPI" {
+		pkgName = strings.ToLower(pkgName)
+	}
+
+	return fmt.Sprintf("https://deps.dev/%s/%s", system, url.PathEscape(pkgName))
+}
+
+// EcosystemLink generates the external package repository link for the affected package.
+func (p AffectedPackageDisplay) EcosystemLink() string {
+	if p.EcosystemName == "" || p.PackageName == "" {
+		return ""
+	}
+
+	baseEco, _, _ := strings.Cut(p.EcosystemName, ":")
+	baseURL, exists := ecosystemPackageURLs[baseEco]
+	if !exists {
+		return ""
+	}
+
+	return baseURL + p.PackageName
+}
+
+// DatabaseSpecificJSON returns the formatted JSON string for the root vulnerability database_specific field.
+func (v VulnerabilityPageData) DatabaseSpecificJSON() string {
+	if v.Vulnerability == nil {
+		return ""
+	}
+
+	return FormatProtoStruct(v.Vulnerability.GetDatabaseSpecific())
+}
+
+// FormattedPublished returns the published timestamp formatted as RFC3339 UTC string.
+func (v VulnerabilityPageData) FormattedPublished() string {
+	if v.Vulnerability.GetPublished() == nil {
+		return ""
+	}
+
+	return v.Vulnerability.GetPublished().AsTime().Format(time.RFC3339)
+}
+
+// FormattedModified returns the modified timestamp formatted as RFC3339 UTC string.
+func (v VulnerabilityPageData) FormattedModified() string {
+	if v.Vulnerability.GetModified() == nil {
+		return ""
+	}
+
+	return v.Vulnerability.GetModified().AsTime().Format(time.RFC3339)
+}
+
+// FormattedWithdrawn returns the withdrawn timestamp formatted as RFC3339 UTC string, or empty string if not withdrawn.
+func (v VulnerabilityPageData) FormattedWithdrawn() string {
+	if v.Vulnerability.GetWithdrawn() == nil {
+		return ""
+	}
+
+	return v.Vulnerability.GetWithdrawn().AsTime().Format(time.RFC3339)
+}
+
+// SummaryHTML returns HTML-escaped summary text with literal backticks wrapped in <span class="literal-backtick">`</span>.
+func (v VulnerabilityPageData) SummaryHTML() template.HTML {
+	if v.Vulnerability == nil || v.Vulnerability.GetSummary() == "" {
+		return ""
+	}
+
+	summary := v.Vulnerability.GetSummary()
+	escaped := template.HTMLEscapeString(summary)
+	if !strings.Contains(escaped, "`") {
+		//nolint:gosec // HTML escaped string without backticks
+		return template.HTML(escaped)
+	}
+
+	replaced := strings.ReplaceAll(escaped, "`", `<span class="literal-backtick">`+"`"+`</span>`)
+	//nolint:gosec // HTML escaped with backtick spans
+	return template.HTML(replaced)
+}
+
+// DetailsHTML returns the details string converted to template.HTML.
+func (v VulnerabilityPageData) DetailsHTML() template.HTML {
+	if v.Vulnerability == nil {
+		return ""
+	}
+
+	return renderMarkdownToHTML(v.Vulnerability.GetDetails())
+}
+
+// IsKnownID checks whether an alias or related vulnerability ID is recognized in the database.
+func (v VulnerabilityPageData) IsKnownID(id string) bool {
+	if v.KnownIDs == nil {
+		return false
+	}
+
+	return v.KnownIDs[id]
+}
+
+// ShouldCollapse returns true if the affected packages list should collapse into grouped accordions.
+func (v VulnerabilityPageData) ShouldCollapse() bool {
+	if v.Vulnerability == nil {
+		return false
+	}
+
+	affected := v.Vulnerability.GetAffected()
+	if len(affected) > 5 {
+		return true
+	}
+
+	var totalEcosystemLen, totalPackageLen int
+	for _, entry := range affected {
+		if pkg := entry.GetPackage(); pkg != nil {
+			totalEcosystemLen += len(pkg.GetEcosystem())
+			totalPackageLen += len(pkg.GetName())
+		}
+		if ranges := entry.GetRanges(); len(ranges) > 0 {
+			totalPackageLen += len(ranges[0].GetRepo())
+		}
+	}
+
+	return max(totalEcosystemLen, totalPackageLen) > 100
+}
+
+// Severities calculates and returns display data for all severity entries on the vulnerability.
+func (v VulnerabilityPageData) Severities() []SeverityDisplay {
+	if v.Vulnerability == nil {
+		return nil
+	}
+
+	severities := v.Vulnerability.GetSeverity()
+	if len(severities) == 0 {
+		return nil
+	}
+
+	id := v.Vulnerability.GetId()
+	displays := make([]SeverityDisplay, 0, len(severities))
+	for _, sev := range severities {
+		score := sev.GetScore()
+		var isCVSS bool
+		var cvssScore float64
+		var cvssLevel string
+		var calcURL string
+		var rating string
+		switch sev.GetType() {
+		case osvschema.Severity_CVSS_V2:
+			isCVSS = true
+			parsed, err := gocvss20.ParseVector(score)
+			if err != nil {
+				logger.Error("failed to parse CVSS V2 vector", slog.String("id", id), slog.Any("error", err))
+				continue
+			}
+			cvssScore = parsed.BaseScore()
+			// CVSS:2.0 does not define ratings
+			calcURL = firstCVSSCalculatorBaseURL + "/2.0#" + score
+		case osvschema.Severity_CVSS_V3:
+			isCVSS = true
+			if strings.HasPrefix(score, "CVSS:3.0") {
+				parsed, err := gocvss30.ParseVector(score)
+				if err != nil {
+					logger.Error("failed to parse CVSS V3.0 vector", slog.String("id", id), slog.Any("error", err))
+					continue
+				}
+				cvssScore = parsed.BaseScore()
+				cvssLevel, _ = gocvss30.Rating(cvssScore)
+				calcURL = firstCVSSCalculatorBaseURL + "/3.0#" + score
+			} else if strings.HasPrefix(score, "CVSS:3.1") {
+				parsed, err := gocvss31.ParseVector(score)
+				if err != nil {
+					logger.Error("failed to parse CVSS V3.1 vector", slog.String("id", id), slog.Any("error", err))
+					continue
+				}
+				cvssScore = parsed.BaseScore()
+				cvssLevel, _ = gocvss31.Rating(cvssScore)
+				calcURL = firstCVSSCalculatorBaseURL + "/3.1#" + score
+			} else {
+				logger.Error("failed to parse CVSS V3 vector: unsupported version", slog.String("id", id), slog.String("score", score))
+				continue
+			}
+		case osvschema.Severity_CVSS_V4:
+			isCVSS = true
+			parsed, err := gocvss40.ParseVector(score)
+			if err != nil {
+				logger.Error("failed to parse CVSS V4.0 vector", slog.String("id", id), slog.Any("error", err))
+				continue
+			}
+			cvssScore = parsed.Score()
+			cvssLevel, _ = gocvss40.Rating(cvssScore)
+			calcURL = firstCVSSCalculatorBaseURL + "/4.0#" + score
+		case osvschema.Severity_Ubuntu:
+		default:
+			logger.Error("unknown severity type", slog.String("id", id), slog.Any("type", sev.GetType()))
+
+			continue
+		}
+		if isCVSS {
+			if cvssLevel != "" {
+				rating = fmt.Sprintf("%.1f (%s)", cvssScore, cases.Title(language.English).String(cvssLevel))
+			} else {
+				rating = fmt.Sprintf("%.1f", cvssScore)
+			}
+		}
+
+		displays = append(displays, SeverityDisplay{
+			Type:          sev.GetType().String(),
+			Score:         score,
+			IsCVSS:        isCVSS,
+			Level:         strings.ToLower(cvssLevel),
+			Rating:        rating,
+			CalculatorURL: calcURL,
+		})
+	}
+
+	return displays
+}
+
+// AffectedPackages parses and returns display objects for all affected packages.
+func (v VulnerabilityPageData) AffectedPackages() []AffectedPackageDisplay {
+	if v.Vulnerability == nil {
+		return nil
+	}
+
+	affectedList := v.Vulnerability.GetAffected()
+	if len(affectedList) == 0 {
+		return nil
+	}
+
+	packages := make([]AffectedPackageDisplay, 0, len(affectedList))
+	for i, aff := range affectedList {
+		ecoName := "Git"
+		pkgName := ""
+		pkgPurl := ""
+		hasPkg := aff.GetPackage() != nil
+
+		if pkg := aff.GetPackage(); pkg != nil {
+			ecoName = pkg.GetEcosystem()
+			pkgName = pkg.GetName()
+			pkgPurl = pkg.GetPurl()
+		}
+
+		// Convert ranges
+		ranges := make([]AffectedRangeDisplay, 0, len(aff.GetRanges()))
+		for _, r := range aff.GetRanges() {
+			repoURL := r.GetRepo()
+			events := make([]AffectedEventDisplay, 0, len(r.GetEvents()))
+			for _, e := range r.GetEvents() {
+				if val := e.GetIntroduced(); val != "" {
+					disp := AffectedEventDisplay{
+						Type:      "Introduced",
+						Value:     val,
+						Repo:      repoURL,
+						IsUnknown: val == "0",
+					}
+					if disp.IsUnknown {
+						if r.GetType() == osvschema.Range_GIT {
+							disp.TooltipText = "Unknown introduced commit / All previous commits are affected"
+						} else {
+							disp.TooltipText = "Unknown introduced version / All previous versions are affected"
+						}
+					}
+					events = append(events, disp)
+				}
+				if val := e.GetFixed(); val != "" {
+					events = append(events, AffectedEventDisplay{
+						Type:  "Fixed",
+						Value: val,
+						Repo:  repoURL,
+					})
+				}
+				if val := e.GetLastAffected(); val != "" {
+					events = append(events, AffectedEventDisplay{
+						Type:  "Last Affected",
+						Value: val,
+						Repo:  repoURL,
+					})
+				}
+				if val := e.GetLimit(); val != "" {
+					events = append(events, AffectedEventDisplay{
+						Type:  "Limit",
+						Value: val,
+						Repo:  repoURL,
+					})
+				}
+			}
+
+			ranges = append(ranges, AffectedRangeDisplay{
+				Type:                 r.GetType().String(),
+				Repo:                 repoURL,
+				Events:               events,
+				DatabaseSpecificJSON: FormatProtoStruct(r.GetDatabaseSpecific()),
+			})
+		}
+
+		if !hasPkg && pkgName == "" {
+			for _, r := range ranges {
+				if r.Repo != "" {
+					pkgName = StripScheme(r.Repo)
+					break
+				}
+			}
+		}
+
+		var versionGroups []VersionGroupDisplay
+		if len(aff.GetVersions()) > 0 {
+			versionGroups = GroupVersions(aff.GetVersions(), ecoName)
+		}
+
+		packages = append(packages, AffectedPackageDisplay{
+			EcosystemName:         ecoName,
+			PackageName:           pkgName,
+			PackagePurl:           pkgPurl,
+			HasPackage:            hasPkg,
+			IsLastPackage:         i == len(affectedList)-1,
+			Ranges:                ranges,
+			VersionGroups:         versionGroups,
+			EcosystemSpecificJSON: FormatProtoStruct(aff.GetEcosystemSpecific()),
+			DatabaseSpecificJSON:  FormatProtoStruct(aff.GetDatabaseSpecific()),
+			DatabaseSpecificKVs:   ParseDatabaseSpecificKVs(aff.GetDatabaseSpecific()),
+		})
+	}
+
+	return packages
+}
+
+// EcosystemGroups groups affected packages by ecosystem for accordion rendering.
+func (v VulnerabilityPageData) EcosystemGroups() []AffectedEcosystemGroup {
+	allPackages := v.AffectedPackages()
+	if len(allPackages) == 0 {
+		return nil
+	}
+
+	grouped := make(map[string][]AffectedPackageDisplay)
+	for _, pkg := range allPackages {
+		eco := pkg.EcosystemName
+		grouped[eco] = append(grouped[eco], pkg)
+	}
+
+	order := make([]string, 0, len(grouped))
+	for eco := range grouped {
+		order = append(order, eco)
+	}
+
+	// Sort ecosystems alphabetically, placing 'Git' at the end
+	slices.SortFunc(order, func(a, b string) int {
+		aIsGit := strings.EqualFold(a, "git")
+		bIsGit := strings.EqualFold(b, "git")
+		if aIsGit && !bIsGit {
+			return 1
+		}
+		if !aIsGit && bIsGit {
+			return -1
+		}
+
+		return strings.Compare(strings.ToLower(a), strings.ToLower(b))
+	})
+
+	groups := make([]AffectedEcosystemGroup, 0, len(order))
+	for i, eco := range order {
+		pkgs := grouped[eco]
+		groups = append(groups, AffectedEcosystemGroup{
+			EcosystemName: eco,
+			IsLast:        i == len(order)-1,
+			Packages:      pkgs,
+		})
+	}
+
+	return groups
+}
