+++ go/internal/website/vulnerability_view_test.go
@@ -1,0 +1,265 @@
+package website
+
+import (
+	"testing"
+
+	"github.com/ossf/osv-schema/bindings/go/osvschema"
+)
+
+func TestSeverities(t *testing.T) {
+	tests := []struct {
+		name        string
+		sevType     osvschema.Severity_Type
+		score       string
+		wantCalcURL string
+		wantLevel   string
+		wantCount   int
+	}{
+		{
+			name:        "CVSS v2.0 valid vector",
+			sevType:     osvschema.Severity_CVSS_V2,
+			score:       "AV:N/AC:L/Au:N/C:P/I:P/A:P",
+			wantCalcURL: firstCVSSCalculatorBaseURL + "/2.0#AV:N/AC:L/Au:N/C:P/I:P/A:P",
+			wantLevel:   "",
+			wantCount:   1,
+		},
+		{
+			name:        "CVSS v3.0 valid vector",
+			sevType:     osvschema.Severity_CVSS_V3,
+			score:       "CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+			wantCalcURL: firstCVSSCalculatorBaseURL + "/3.0#CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+			wantLevel:   "critical",
+			wantCount:   1,
+		},
+		{
+			name:        "CVSS v3.1 valid vector",
+			sevType:     osvschema.Severity_CVSS_V3,
+			score:       "CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+			wantCalcURL: firstCVSSCalculatorBaseURL + "/3.1#CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+			wantLevel:   "critical",
+			wantCount:   1,
+		},
+		{
+			name:        "CVSS v4.0 valid vector",
+			sevType:     osvschema.Severity_CVSS_V4,
+			score:       "CVSS:4.0/AV:N/AC:L/AT:N/PR:N/UI:N/VC:H/VI:H/VA:H/SC:N/SI:N/SA:N",
+			wantCalcURL: firstCVSSCalculatorBaseURL + "/4.0#CVSS:4.0/AV:N/AC:L/AT:N/PR:N/UI:N/VC:H/VI:H/VA:H/SC:N/SI:N/SA:N",
+			wantLevel:   "critical",
+			wantCount:   1,
+		},
+		{
+			name:      "CVSS v3 unsupported/invalid prefix",
+			sevType:   osvschema.Severity_CVSS_V3,
+			score:     "CVSS:3.2/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+			wantCount: 0,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pageData := VulnerabilityPageData{
+				Vulnerability: &osvschema.Vulnerability{
+					Id: "GHSA-1234-abcd",
+					Severity: []*osvschema.Severity{
+						{
+							Type:  tt.sevType,
+							Score: tt.score,
+						},
+					},
+				},
+			}
+
+			displays := pageData.Severities()
+			if len(displays) != tt.wantCount {
+				t.Fatalf("Severities() count = %d, want %d", len(displays), tt.wantCount)
+			}
+			if tt.wantCount > 0 {
+				d := displays[0]
+				if d.CalculatorURL != tt.wantCalcURL {
+					t.Errorf("CalculatorURL = %q, want %q", d.CalculatorURL, tt.wantCalcURL)
+				}
+				if d.Level != tt.wantLevel {
+					t.Errorf("Level = %q, want %q", d.Level, tt.wantLevel)
+				}
+			}
+		})
+	}
+}
+
+func TestDepsDevURL(t *testing.T) {
+	tests := []struct {
+		name      string
+		ecosystem string
+		pkgName   string
+		wantURL   string
+	}{
+		{
+			name:      "PyPI package lowercased",
+			ecosystem: "PyPI",
+			pkgName:   "Requests",
+			wantURL:   "https://deps.dev/pypi/requests",
+		},
+		{
+			name:      "npm package",
+			ecosystem: "npm",
+			pkgName:   "express",
+			wantURL:   "https://deps.dev/npm/express",
+		},
+		{
+			name:      "Go package path escaped",
+			ecosystem: "Go",
+			pkgName:   "github.com/gin-gonic/gin",
+			wantURL:   "https://deps.dev/go/github.com%2Fgin-gonic%2Fgin",
+		},
+		{
+			name:      "Ecosystem with colon qualifier",
+			ecosystem: "PyPI:SubPackage",
+			pkgName:   "Flask",
+			wantURL:   "https://deps.dev/pypi/flask",
+		},
+		{
+			name:      "Unknown ecosystem returns empty string",
+			ecosystem: "UnknownEco",
+			pkgName:   "some-pkg",
+			wantURL:   "",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pkg := AffectedPackageDisplay{
+				EcosystemName: tt.ecosystem,
+				PackageName:   tt.pkgName,
+			}
+			got := pkg.DepsDevURL()
+			if got != tt.wantURL {
+				t.Errorf("DepsDevURL() = %q, want %q", got, tt.wantURL)
+			}
+		})
+	}
+}
+
+func TestEcosystemLink(t *testing.T) {
+	tests := []struct {
+		name      string
+		ecosystem string
+		pkgName   string
+		wantURL   string
+	}{
+		{
+			name:      "PyPI package link",
+			ecosystem: "PyPI",
+			pkgName:   "requests",
+			wantURL:   "https://pypi.org/project/requests",
+		},
+		{
+			name:      "npm package link",
+			ecosystem: "npm",
+			pkgName:   "react",
+			wantURL:   "https://www.npmjs.com/package/react",
+		},
+		{
+			name:      "crates.io package link",
+			ecosystem: "crates.io",
+			pkgName:   "tokio",
+			wantURL:   "https://crates.io/crates/tokio",
+		},
+		{
+			name:      "Unsupported ecosystem",
+			ecosystem: "UnknownEco",
+			pkgName:   "foo",
+			wantURL:   "",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pkg := AffectedPackageDisplay{
+				EcosystemName: tt.ecosystem,
+				PackageName:   tt.pkgName,
+			}
+			got := pkg.EcosystemLink()
+			if got != tt.wantURL {
+				t.Errorf("EcosystemLink() = %q, want %q", got, tt.wantURL)
+			}
+		})
+	}
+}
+
+func TestSummaryHTML(t *testing.T) {
+	tests := []struct {
+		name    string
+		summary string
+		want    string
+	}{
+		{
+			name:    "Empty summary",
+			summary: "",
+			want:    "",
+		},
+		{
+			name:    "Plain summary HTML escaped",
+			summary: "Buffer overflow in <script>alert(1)</script>",
+			want:    "Buffer overflow in &lt;script&gt;alert(1)&lt;/script&gt;",
+		},
+		{
+			name:    "Summary with backticks",
+			summary: "Crash in `foo()` function",
+			want:    `Crash in <span class="literal-backtick">` + "`" + `</span>foo()<span class="literal-backtick">` + "`" + `</span> function`,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pageData := VulnerabilityPageData{
+				Vulnerability: &osvschema.Vulnerability{
+					Summary: tt.summary,
+				},
+			}
+			got := string(pageData.SummaryHTML())
+			if got != tt.want {
+				t.Errorf("SummaryHTML() = %q, want %q", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestEcosystemGroups(t *testing.T) {
+	pageData := VulnerabilityPageData{
+		Vulnerability: &osvschema.Vulnerability{
+			Affected: []*osvschema.Affected{
+				{
+					Package: &osvschema.Package{
+						Ecosystem: "Git",
+						Name:      "github.com/example/repo",
+					},
+				},
+				{
+					Package: &osvschema.Package{
+						Ecosystem: "PyPI",
+						Name:      "requests",
+					},
+				},
+				{
+					Package: &osvschema.Package{
+						Ecosystem: "npm",
+						Name:      "express",
+					},
+				},
+			},
+		},
+	}
+
+	groups := pageData.EcosystemGroups()
+	if len(groups) != 3 {
+		t.Fatalf("EcosystemGroups() count = %d, want 3", len(groups))
+	}
+
+	// Should be sorted alphabetically (npm, PyPI) with 'Git' placed last.
+	wantOrder := []string{"npm", "PyPI", "Git"}
+	for i, want := range wantOrder {
+		if groups[i].EcosystemName != want {
+			t.Errorf("EcosystemGroups()[%d] = %q, want %q", i, groups[i].EcosystemName, want)
+		}
+	}
+}
